output log in run tasks
